chore: landing 폴더 interface, style 파일 분리
