+++ src/Page/Landing/interface.ts
@@ -1,0 +1,16 @@
+export interface SignUpProps {
+  email: string;
+  name: string;
+  password: string;
+  passwordConfirm: string;
+  code: string;
+  storeName: string;
+  address: string;
+  phone: string;
+}
+export interface StoreProps {
+  code: string;
+  name: string;
+  address: string;
+  phone: string;
+}
